+++ src/uio/Edit/Value.h
@@ -1,4 +1,4 @@
-//  Copyright (c) 2018 Hugo Amiard hugo.amiard@laposte.net
+//  Copyright (c) 2019 Hugo Amiard hugo.amiard@laposte.net
 //  This software is provided 'as-is' under the zlib License, see the LICENSE.txt file.
 //  This notice and the license may not be removed or altered from any source distribution.
 
@@ -6,7 +6,7 @@
 
 #ifndef MUD_MODULES
 #include <infra/Global.h>
-#include <obj/Dispatch.h>
+#include <type/Dispatch.h>
 #include <math/Stat.h>
 #endif
 #include <uio/Forward.h>
@@ -20,7 +20,7 @@ namespace mud
 		DispatchInput();
 	};
 
-	export_ MUD_UIO_EXPORT bool type_selector(Widget& parent, size_t& type, array<Type*> types);
+	export_ MUD_UIO_EXPORT bool type_selector(Widget& parent, uint32_t& type, array<Type*> types);
 
 	export_ MUD_UIO_EXPORT bool value_edit(Widget& parent, Ref& value);
 	export_ MUD_UIO_EXPORT bool none_edit(Widget& parent, Ref& value);
